Rewrite tap tests with luatest
